Separated entities and forms part 1
