+++ backend/controllers/SiteController.php
@@ -4,7 +4,7 @@ namespace backend\controllers;
 use Yii;
 use yii\web\Controller;
 use yii\filters\VerbFilter;
-use common\models\LoginForm;
+use common\forms\LoginForm;
 
 /**
  * Site controller
