--- backend/models/.gitkeep
@@ -1,1 +1,0 @@
-*
