+++ common/fixtures/UserFixture.php
@@ -5,5 +5,5 @@ use yii\test\ActiveFixture;
 
 class UserFixture extends ActiveFixture
 {
-    public $modelClass = 'common\models\User';
+    public $modelClass = 'common\entities\User';
 }
