+++ common/forms/LoginForm.php
@@ -1,6 +1,7 @@
 <?php
-namespace common\models;
+namespace common\forms;
 
+use common\entities\User;
 use Yii;
 use yii\base\Model;
