+++ common/tests/unit/forms/LoginFormTest.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace common\tests\unit\models;
+namespace common\tests\unit\forms;
 
 use Yii;
-use common\models\LoginForm;
+use common\forms\LoginForm;
 use common\fixtures\UserFixture;
 
 /**
