+++ frontend/config/main.php
@@ -7,20 +7,18 @@ $params = array_merge(
 );
 
 return [
-    'id' => 'app-frontend',
     'basePath' => dirname(__DIR__),
     'bootstrap' => [
         'common\bootstrap\SetUp',
         'log',
     ],
-    'controllerNamespace' => 'frontend\controllers',
     'components' => [
         'request' => [
             'cookieValidationKey' => $params['cookieValidationKey'],
             'csrfParam' => '_csrf-frontend',
         ],
         'user' => [
-            'identityClass' => 'common\models\User',
+            'identityClass' => 'common\entities\User',
             'enableAutoLogin' => true,
             'identityCookie' => [
                 'name' => '_identity',
@@ -53,5 +51,7 @@ return [
             return Yii::$app->get('frontendUrlManager');
         },
     ],
+    'controllerNamespace' => 'frontend\controllers',
+    'id' => 'app-frontend',
     'params' => $params,
 ];
